requirements.txt
selenium
psycopg2-binary
requests